+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # mu-player
+
+### Changelog
+    - Playlist: remove tracks
+    - Track title capitalize
     - No more manual halted searches: timeouts and retries added. Tweak it in cinfig file.
     - Top albums search
-### Changelog
+
   * v0.4.0
     - Better soundcloud artist detection
     - Search code total rewrite: smart logic was implemented